cell-quant-agentic/requirements.txt
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.0.0
openslide-python>=1.1.2
python-bioformats>=4.0.0
javabridge>=1.0.19

# VLM dependencies
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
sentencepiece>=0.1.99
protobuf>=3.20.0
langchain>=0.0.200
langchain-community>=0.0.10
llama-cpp-python>=0.2.6
safetensors>=0.3.1

# VLM-specific packages
git+https://github.com/IDEA-Research/GroundingDINO.git
git+https://github.com/facebookresearch/segment-anything.git
ultralytics>=8.0.0  # for YOLOv8
segment-anything>=1.0.0
medsam>=0.1.0
seggpt>=0.1.0

# Additional utilities
tqdm>=4.65.0
wandb>=0.15.0 